requirements.txt
Flask==1.1.2
wtforms==2.3.1
passlib==1.7.2
psycopg2==2.8.5
names==0.3.0
python-dotenv==0.14.0